Interface LIBORMarketModelInterface replaced by narrower interface.
From the interface `LIBORMarketModelInterface` the narrower interface
`LIBORModelInterface` has been extracted,
leaving the methods related to the covariance model to
`LIBORMarketModelInterface` only. The method `getModel()`
of `LIBORModelMonteCarloSimulationInterface` now returns a
`LIBORModelInterface` interface only. In case your
code requires access to the covariance model, you have to check for
`LIBORMarketModelInterface` (see the code
of `getValue` in `SwaptionAnalyticApproximation` for an example).

git-svn-id: http://svn.finmath.net/finmath%20lib/trunk@1783 48fcbb9e-5639-0410-8b25-842128a4905d