Refactoring.
The objects implementing AbstractLIBORCovarianceModelParametric have
become immutable. The method setParameter has been removed and replaced
by a method getCloneWithModifiedParameter. This allows a performance
increase, since in a calibration it is now safe to reuse parts of the
model, given that parts of a parameter (e.g. correlation) has not
changed.