The changes of Deposit try to fix three things:
1) In the current implementation a deposit set up with the current
market rate (as calculated by getRate) does not have getValue=0. This is
because the rate calculation was slightly flawed (line 92).

2) The value of a deposit should be the present value of the payout
cashflow (which I would expect to be at start date) and the payback
cashflow at maturity. This was not the case. 

3) The getValue function did not take into account the argument
evaluationTime. The implementation provided in this change, asserts that
the value after maturity is zero and neglects the payout cashflow once
it is in the past.

These three things are checked by the DepositTest class.