General Overhaul of Forward Curves
- ForwardCurveFromDiscountCurve: Added member
referenceDiscountCurveForForwardsName which is the (pseudo-)discount
curve that the forwards are calculated from. Note that this is in
general different from the discount curve associated with the forwards
- ForwardCurveInterface: New method getForward(double fixingTime)
without a model. This returns the forward whenever no model is required
for the calculation and throws an exception otherwise 
- AbstractForwardCurve/ForwardCurve/ForwardCurveFromDiscountCurve:
Reduced code duplication by letting constructors call each other
- ForwardCurveFromDiscountCurve: Added to String Method