Fixed bug induced by toString comparison and improved performance.
The comparisson of the dates via toString is flawed since toString
prints the LocalDate in ISO-8601 (or maybe even in Locale dependent
format?).

In addition: the repetitive call to toList is slow and contains in
ArrayList is of O(n) instead of O(log(n)). This has been improved using
static initialized and TreeSet.