Refactoring
- The interface net.finmath.modelling.ModelInterface has been
renamed/refactored to net.finmath.modelling.Model<T extends
ModelDescriptor>. This is related to the introduction of
ModelDescriptors and ProductDescriptors. The interface was previously a
marker interface, but now comes with two methods.
- The interface net.finmath.modelling.ProductInterface has been
renamed/refactored to net.finmath.modelling.Product<T extends
ProductDescriptor>. This is related to the introduction of
ModelDescriptors and ProductDescriptors. The interface was previously a
marker interface, but now comes with two methods.
- Classes implementing the interface Product<T> require a type
parameterization, example: ```AbstractMonteCarloProduct<T extends
ProductDescriptor> implements Product<T>'''