Refactoring. Removing product assumptions from solver.
Extracted the interface FiniteDifference1DBoundary which is provided by
a product to the model and gives the boundary conditional.

Letting FDMEuropeanCallOption implements FiniteDifference1DBoundary

Renamed expectedValueOfStockPrice to getForwardValue - not sure if this
method will stay there.

Removed dependency on option by making „center“ (strike) and
"timeHorizon" (maturity) solver parameters

Renamed the method valueOption to getValue. - because the solver is
(will be) not only for options.

In FDMThetaMethod the member option is now renamed. It is also a
  FiniteDifference1DBoundary boundaryCondition