Refactoring of model and product interfaces.
The interface providing model and product descriptors is now separated
from the model interface and the product interfaces. This allows one for
example one product to implement different product interfaces (e.g. for
different numerical methods) and implementing one constructor for a
product descriptor. The relation product descriptor to implementations
is 1:n.