Backward compatible refactoring of ProcessModel.
The method getNumeraire, getDrift and getFactorLoading of the model
(ProcessModel) now require the process as an argument.

The methods without this argument are provided for backward compatiblity
in this version.

In future versions ProcessModel no longer caries a reference to the
process in order to be applicable for different processes.

The only strong implications are relatied to the implementation of
caches.