Interest rate Monte-Carlo simulations implement the slightly weaker interface TermStructureMonteCarloSimulationModel
The class TermStructureMonteCarloSimulationFromTermStructureModel
implements the weaker interface
TermStructureMonteCarloSimulationFromTermStructureModel. This is used in
valuation of interest rate products. This interface does not provide the
methods related to forward rate period discretizations -
getLIBORPeriodDiscretization.

In case it is needed, there is still a
LIBORMonteCarloSimulationFromTermStructureModel implementing
LIBORModelMonteCarloSimulationModel for backward compatibility (however,
here the corresponding methods thrown an exception). (This was the case
in the previous version).

As before, the class LIBORMonteCarloSimulationFromLIBORModel implements
LIBORModelMonteCarloSimulationModel and provides a
getLIBORPeriodDiscretization().